Implementing error message for empty forms before edting

+++ database/migrations/2020_06_11_120438_create_patient__breath_alcohol_test_and_electrocardiograms_table.php
@@ -19,7 +19,7 @@ class CreatePatientBreathAlcoholTestAndElectrocardiogramsTable extends Migration
             $table->date('dateTaken');
             $table->time('timeTaken');
             $table->string('laboratory');
-            $table->integer('breathalcohol');
+            $table->decimal('breathalcohol');
             $table->string('breathalcoholResult');
             $table->string('Usertranscribed');
             //ECG starts from here
